% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_hexagons}
\alias{example_dataset_hexagons}
\title{Hexagon data}
\format{
An sf data frame with 1953 rows and 2 columns:
\describe{
  \item{target_id}{The name of the hexagon.}
  \item{geometry}{An sf column giving the boundaries of the hexagon.}
}
}
\source{
Gabriel Ahlfeldt,
}
\usage{
example_dataset_hexagons
}
\description{
A shapefile of various hexagons covering Berlin.
}
\keyword{datasets}
