% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_transactions}
\alias{example_dataset_transactions}
\title{Transaction data}
\format{
An sf data frame with 481,159 rows and 5 columns:
\describe{
  \item{price}{Price of transaction in pounds.}
  \item{year}{Year of transaction}
  \item{Att_newbuilt}{A dummy for whether the property is newbuilt}
  \item{submarket}{Submarket 2 is all properties south of the Thames, submarket 1 is all properties north of the Thames}
  \item{geometry}{An sf column giving the location of the property as a point.}
}
}
\source{
HM Land Registry's Price Paid Data (\url{https://www.gov.uk/government/collections/price-paid-data}). Ordnance Survey's Code Point open (\url{https://www.data.gov.uk/dataset/c1e0176d-59fb-4a8c-92c9-c8b376a80687/code-point-open})
}
\usage{
example_dataset_transactions
}
\description{
Geolocated property transactions in Greater London in 2015 to 2018.
}
\keyword{datasets}
