% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_wards}
\alias{example_dataset_wards}
\title{Ward data}
\format{
An sf data frame with 253 rows and 4 columns:
\describe{
  \item{NAME}{The name of the ward.}
  \item{target_id}{The ID of the ward.}
  \item{DISTRICT}{The district to which the ward belongs.}
  \item{geometry}{An sf column giving the boundaries of the ward}
}
}
\source{
Greater London Authority's DataStore, manipulated by Alexander Hansen. Open Govenrment License v2. \url{https://data.london.gov.uk/dataset/statistical-gis-boundary-files-london}
}
\usage{
example_dataset_wards
}
\description{
A shapefile of wards in Inner London
}
\keyword{datasets}
