% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{transform_crs_to_utm}
\alias{transform_crs_to_utm}
\title{Transform CRS into correct UTM.}
\usage{
transform_crs_to_utm(sf_df, debug = T)
}
\arguments{
\item{sf_df}{An sf dataframe to be transformed}

\item{debug}{Whether to print debugging messages. Defaults to TRUE}
}
\value{
An sf dataframe in the correct UTM format.
}
\description{
Transform CRS into correct UTM.
}
\examples{
#sf_df <- read_sf(xxx)
#sf_sf <- transform_crs_to_utm(sf_sf)
}
