% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_index}
\alias{calculate_index}
\title{Calculate house price index}
\usage{
calculate_index(
  target_dataset,
  transaction_dataset,
  observations_outer,
  observations_inner,
  max_radius_outer = 20,
  max_radius_inner = 10,
  n_cores = NULL,
  use_parallel = T,
  skip_errors = F,
  allow_expansion = T,
  debug = F
)
}
\arguments{
\item{target_dataset}{The dataset with all targets. Each target must be a separate row.
The dataset must be either an sf dataframe or a dataframe with 'target_X' and 'target_Y' columns in projected meters.
It must have a target identifier named 'target_id'.}

\item{transaction_dataset}{The dataset with all transactions. Each transaction must be a separate row.
The dataset must be either an sf dataframe or a dataframe with 'origin_X' and 'origin_Y' columns in projected meters.
The dataset must have columns named 'year', 'submarket' and 'price'. Any attributes to be included in the regression as hedonics must be called "Att_xxx", with xxx being replaceable.}

\item{observations_outer}{The target number of transactions for the outer ring.}

\item{observations_inner}{The target number of transactions for the inner ring.}

\item{max_radius_outer}{The maximal radius in kilometers to be taken by the outer ring. Default value 20 km.}

\item{max_radius_inner}{The maximal radius in kilometers to be taken by the inner ring. Default value 10 km.}

\item{n_cores}{The number of cores to be used. Default is NULL, meaning that the program will use all but one available cores.}

\item{use_parallel}{Whether to use parallel processing. Defaults to TRUE. Turning off can be useful for debugging.}

\item{skip_errors}{Whether to skip problematic targets. Defaults to FALSE, meaning an error will force the function to stop. Changing this parameter makes debugging very hard.}

\item{allow_expansion}{Whether to allow number of included observations to exceed observations_outer param. Defaults to TRUE, meaning radius will increase for sparse target-year combinations.}

\item{debug}{Whether to print all messages for debugging purpose. Defaults to FALSE.}
}
\value{
A single data.table with a row for each year-target pair.
}
\description{
The function uses a dataset with transactions and a dataset with a list of targets where
the index is to be calculated.
}
\examples{
#transactions <- read_sf(xxx)
#targets <- read_sf(yyy)
#index = calculate_index(targets, transactions, observations_outer=10000, observations_inner=2000)
}
