% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{calculate_index_point}
\alias{calculate_index_point}
\title{Calculate index for a single target point in space}
\usage{
calculate_index_point(
  one_target,
  list_of_transactions,
  A_N,
  T_N,
  max_radius_A,
  max_radius_T,
  treatment_vars,
  debug,
  N_year
)
}
\arguments{
\item{one_target}{The target for which an index is to be calculated}

\item{list_of_transactions}{The dataset with all transactions.}

\item{A_N}{The target number of transactions for the outer ring.}

\item{T_N}{The target number of transactions for the inner ring.}

\item{max_radius_A}{The maximal radius in kilometers to be taken by the outer ring.}

\item{max_radius_T}{The maximal radius in kilometers to be taken by the inner ring.}

\item{treatment_vars}{List of variables containing Attributes to be included in regression.}

\item{debug}{Whether to print messages for debugging.}

\item{N_year}{Number of years to calculate the index for.}
}
\value{
A data.table
}
\description{
Calculate index for a single target point in space
}
\examples{
#On the way
}
