% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_dataset_transactions}
\alias{example_dataset_transactions}
\title{Transaction data}
\format{
An sf data frame with 1,648,823 rows and 6 columns:
\describe{
  \item{price}{Price of transaction in euro per square meter.}
  \item{year}{Year of transaction}
  \item{Att_area}{Size of transacted property in square meters.}
  \item{Att_balcon}{A dummy for whether the property has a balcony.}
  \item{submarket}{An indicator for whether the property is in East Berlin, West Berlin or Rest of Germany}
  \item{geometry}{An sf column giving the location of the property as a point.}
}
}
\source{
Gabriel Ahlfeldt,
}
\usage{
example_dataset_transactions
}
\description{
An artificial dataset mimicking property transactions in Berlin.
}
\keyword{datasets}
